Factor out a Routes mixin for Participant
